[#19] Extract screen contents
